twitter-kafka/requirements.txt
tweepy==3.5.0
kafka-python==0.9.5
#pykafka==2.2.0
